Refactoring grades and test
Instead of using two arrays, one for the value of the grade and one for the validation of grades, I changed for one array of object.
This object has the attributes value and valid, which represent the value and validity of the grade.
Also refactored test to confirm with approach.